refactor(main/index.js): modularizando arquivo;
dividindo arquivo index.js em 2 arquivos com funções especializadas para cada arquivo, sendo window.js para configurações de janela e ipcHandlers para procedimentos ipc.
